+++ user_docs/en/changes.md
@@ -23,6 +23,7 @@ The available options are:
 * The `-c`/`--config-path` and `--disable-addons` command line options are now respected when launching an update from within NVDA. (#16937)
 * eSpeak NG has been updated to 1.52-dev commit `961454ff`. (#16775)
   * Added new languages Faroese and Xextan.
+* When using a multi-line braille display via the Standard HID braille driver, all lines of cells will be used. (#16993, @alexmoon)
 
 ### Bug Fixes